fix(runtime): make launcher restarts ownership-safe
fix: preserve live foreign-launcher ownership and fail closed when runtime ownership state cannot be persisted.

fix: negotiate an authenticated atomic drain-if-idle lease so busy browser or HTTP turns keep admission open and uncertain delivery is compensated safely.

test: split the runtime supervisor regression suite below the 500-line limit and cover stale owners, persistence failure, tunnel recovery, drain races, and restart compensation.

chore: extract lifecycle authorization and HTTP turn accounting from the server to keep production modules within the repository line limit.
